refactor the todos logic from the App.vue component 1. separate logic into one file then spread in the App component 2. remove the modelsWrapper.js function, actually this a helper that i follow from some tutorial but did not used at all 3. remove the un used input property in Input Component
