+++ src/App.vue
@@ -1,6 +1,11 @@
 <template>
   <div class="container py-5">
-    <alert :text="errorMessage" v-if="error" />
+    <alert
+      :text="errorMessage"
+      v-if="error"
+      class="animate__animated"
+      :class="animateError"
+    />
     <card>
       <template v-slot:content>
         <h1 class="card-title">Submit A Todo</h1>
@@ -11,19 +16,13 @@
       <card>
         <template v-slot:content>
           <div class="mb-3">
-            <Input
-              label="todo"
-              @input="inputTitle"
-              :value="title"
-              :validate="validateForm"
-            />
+            <Input label="todo" @input="inputTitle" :value="title" />
           </div>
           <div class="mb-3">
             <Input
               label="description"
               @input="inputDesc"
               :value="description"
-              :validate="validateForm"
             />
           </div>
           <div class="mb-3">
@@ -59,125 +58,16 @@ import Card from "@/components/Card.vue";
 import Input from "@/components/Input.vue";
 import Todo from "@/components/Todo.vue";
 import Alert from "@/components/Alert.vue";
-import { ref, reactive, computed, watchEffect } from "vue";
-import { v4 as uuidv4 } from "uuid";
+import { useTodosAction } from "@/utils/todos.js";
 export default {
   name: "App",
   setup() {
-    const title = ref("");
-    const description = ref("");
-    const error = ref(false);
-    const errorMessage = ref("");
-    const selected = reactive({
-      data: {},
-    });
-    const todos = reactive({
-      data: [],
-    });
-
-    const submitData = () => {
-      const todoObject = {
-        id: uuidv4(),
-        title: title.value,
-        status: false,
-        description: description.value,
-        isUpdate: false,
-        animateEffect: "animate__animated animate__bounce"
-      };
-      if (title.value !== "" && description.value !== "") {
-        todos.data = [...todos.data, todoObject];
-        error.value = false;
-        errorMessage.value = "";
-        clearForm();
-      }else if(title.value === ""){
-         error.value = true;
-         errorMessage.value = "Title must be filled";
-      }else if(description.value === ""){
-         error.value = true;
-         errorMessage.value = "Description must be filled";
-      }else if (description.value === "" && title.value === ""){
-         error.value = true;
-         errorMessage.value = "Kedua form harus di isi";
-      }
-    };
-
-    const clearForm = () => {
-      title.value = "";
-      description.value = "";
-    };
-
-    const showData = (id) => {
-      const selectedTodo = todos.data.find((todo) => todo.id === id);
-      selected.data = selectedTodo;
-      selectedTodo.isUpdate = !selectedTodo.isUpdate;
-    };
-
-    const cancel = (id) => {
-      showData(id);
-      selected.data = {};
-    };
-
-    const deleteData = (id) => {
-      const findTodo = todos.data.filter((todo) => todo.id !== id);
-      findTodo.animateEffect = "animate__animated animate__fadeout";
-      todos.data = findTodo;
-      return todos.data;
-    };
-
-    const updateData = (value) => {
-      const findTodo = todos.data.find((todo) => todo.id === value.id);
-      findTodo.title = value.title;
-      findTodo.description = value.description;
-      findTodo.isUpdate = false;
-    };
-
-    const inputTitle = (value) => (title.value = value);
-    const inputDesc = (value) => (description.value = value);
-
-    watchEffect(() => {
-      if (title.value.length > 0 || description.value.length > 0) {
-        error.value = false;
-      }
-    });
-
-    const validateForm = computed(() => {
-      if (title.value.length === 0) {
-        return {
-          type: "message",
-          message: "Title tidak boleh kosong",
-          error: true,
-        };
-      } else if (description.value.length === 0) {
-        return {
-          message: "Description tidak boleh kosong",
-          error: true,
-        };
-      } else {
-        return {
-          message: "",
-          error: false,
-        };
-      }
-    });
     return {
-      Card,
-      Input,
-      Todo,
       Alert,
-      title,
-      description,
-      todos,
-      submitData,
-      deleteData,
-      error,
-      showData,
-      cancel,
-      selected,
-      inputTitle,
-      inputDesc,
-      updateData,
-      validateForm,
-      errorMessage,
+      Card,
+      Todo,
+      Input,
+      ...useTodosAction(),
     };
   },
 };
