--- src/utils/modelsWrapper.js
@@ -1,8 +1,0 @@
-import { computed } from "vue";
-
-export function useModelWrapper(_, emit, name = "value") {
-    return computed({
-        get: () => _.value,
-        set: (value) => emit(`update:${name}`, value),
-    });
-}
